+++ src/test/compilation-failed/abi-global-errors.spec.ts
@@ -6,6 +6,8 @@ describe("abi/global.ts errors", () => {
         __DANGER_resetNodeId();
     });
 
+    // TODO: fixme
+
     itShouldNotCompile({
         testName: "sha256-expects-string-or-slice",
         errorMessage: "sha256 expects string or slice argument",